Separate test module for tetst creating script from string
